Modify the constructor to use a pointer to const object and add a command method.
Change the member variable _i2c to use a pointer to a constant object. Modify the constructor accordingly. Add a method command in the I2C subclass and a convience method for type casting. Use that method to implement the exitSafeStart method. Add a type specifier in the enum SmcG2Command