Add an encoder class for the pololu motor encoder using the ESP32encoder library.
The encoder class should have an init and update method.
The constructor should create a new ESP32Encoder object and expect the gearing,
the cpr and the radious of the wheel from the user.
The init method setup the ESP32 pins for a full quadrature encoder
The update method should update the internal encoder count